+++ src/core/model/document_node.hpp
@@ -89,7 +89,8 @@ public:
     {
         Transformed,    ///< Paint only this, apply transform
         NoTransform,    ///< Paint only this, don't apply transform
-        Recursive       ///< Paint this and children, apply transform
+        Recursive,      ///< Paint this and children, apply transform
+        Render          ///< Recursive, but hide objects maked with render == false
     };
 
     using ReferenceTarget::ReferenceTarget;